flask-exam-checker (7)/flask-exam-checker (5)/flask-exam-checker (10)/flask-exam-checker (8)/flask-exam-checker (6)/flask-exam-checker (2)/flask-exam-checker/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
google-generativeai==0.3.2
python-dotenv==1.0.0
Pillow==10.1.0
openpyxl==3.1.2
